Simplify attachments
No need to store the origin of faces and update the origin
every time an object moves.  Just compute the origin of the
face at the time of the attachment.

Also, make Object.origin read-only. The origin attribute of OpenSCAD
objects is writable, but writing to it does *not* change the origin of
the object. So make it a little more clear to a user that this is the
case.