WIP: Add a Composition object
This is a work in progress. It has bugs.

A composition is a kind of virtual object with delayed instantiation.
Composition objects are composed with a parent object and other
composition objects using Composition.compose(). At the time of
composition, the objects in the composition are created and any
attachments/justifications are resolved.

Compose is useful when you have multiple overlapping objects where you
would like the voids of an object to create voids in the other objects.